Moves all impl files to separate "src" library
