+++ flow/include/follower/follower.h
@@ -12,6 +12,7 @@
 
 // Flow
 #include <flow/captor.h>
+#include <flow/captor_state.h>
 #include <flow/utility/implement_crtp_base.hpp>
 
 namespace flow
@@ -118,6 +119,6 @@ struct is_follower : std::integral_constant<bool, std::is_base_of<Follower<Capto
 }  // namespace flow
 
 // Flow (implementation)
-#include <flow/follower/impl/follower.hpp>
+#include "flow/src/follower/follower.hpp"
 
 #endif  // FLOW_FOLLOWER_FOLLOWER_H
