+++ flow/src/follower/before.hpp
@@ -4,8 +4,8 @@
  *
  * @warning IMPLEMENTATION ONLY: THIS FILE SHOULD NEVER BE INCLUDED DIRECTLY!
  */
-#ifndef FLOW_FOLLOWER_IMPL_BEFORE_HPP
-#define FLOW_FOLLOWER_IMPL_BEFORE_HPP
+#ifndef FLOW_SRC_FOLLOWER_BEFORE_HPP
+#define FLOW_SRC_FOLLOWER_BEFORE_HPP
 
 // C++ Standard Library
 #include <cstdint>
@@ -80,4 +80,4 @@ void Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>::abort_follower_i
 }  // namespace follower
 }  // namespace flow
 
-#endif  // FLOW_FOLLOWER_IMPL_BEFORE_HPP
+#endif  // FLOW_SRC_FOLLOWER_BEFORE_HPP
